refactor(card): Move all cards to one folder
