+++ src/app/card/event-card/event-card.component.css
