+++ src/app/card/notify-card/notify-card.component.css
