+++ src/app/card/notify-card/notify-card.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input } from '@angular/core';
-import { INotification } from '../interfaces';
-import { DataService } from '../data.service';
-import { API } from '../../api';
+import { INotification } from '../../interfaces';
+import { DataService } from '../../data.service';
+import { API } from '../../../api';
 import { Router } from '@angular/router';
 import { MatBottomSheetRef } from '@angular/material/bottom-sheet';
